+++ scripts/cleanup-out.sh
@@ -1,7 +1,0 @@
-cd out
-rm *.html
-rm api*.json
-rm auth*.json
-rm output*.json
-rm nexus*.json
-cd ..