+++ tests/test_init.py
@@ -1,7 +1,6 @@
 """Base tests for aioamazondevices."""
 
 from aioamazondevices.api import (
-    AmazonDevice,
     AmazonEchoApi,
 )
 from aioamazondevices.exceptions import (
@@ -12,7 +11,6 @@ from aioamazondevices.exceptions import (
 
 def test_objects_can_be_imported() -> None:
     """Verify objects exist."""
-    assert type(AmazonDevice)
     assert type(AmazonEchoApi)
     assert type(CannotConnect)
     assert type(CannotAuthenticate)